Halting development on SceneTree/Scene architecture for now.
I am halting development on the SceneTree architecture for now as whilst it is necessary, its not the most important feature right now.
I have bitten of more than I can chew for the moment and I am going to focus on more basic but essential features of the engine for now.